backend/requirements.txt
Django==4.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.5.0
django-cors-headers==4.7.0
psycopg2-binary==2.9.10
Pillow==11.1.0
python-decouple==3.8
drf-nested-routers==0.94.1
PyJWT==2.9.0
pytest==8.0.1
pytest-django==4.8.0
pytest-cov==4.1.0
black==24.2.0
isort==5.13.2
flake8==7.0.0
django-filter==23.5
drf-yasg==1.21.7
gunicorn==22.0.0
whitenoise==6.6.0 